requirements/development.txt
Django==2.0.7
psycopg2
mysqlclient
django-filter
django-widget-tweaks
pinax-messages
python-dateutil
sorl-thumbnail
requests
djangorestframework
django-rest-auth
django-cors-headers

